requirements.txt
flask
flask_sqlalchemy
flask_bcrypt
flask_login
flask_mail
flask_wtf
email_validator
wtforms
pyjwt